+++ examples/google-gemini/src/functions/google-gemini/index.ts
