+++ examples/google-gemini/src/functions/google-gemini/text/generateContent.ts
