+++ examples/google-gemini/src/functions/google-gemini/text/generateContentStream.ts
