+++ examples/google-gemini/src/functions/google-gemini/text/index.ts
