+++ examples/google-gemini/src/functions/google-gemini/types/index.ts
