+++ examples/google-gemini/src/functions/google-gemini/types/inputs.ts
