+++ examples/google-gemini/src/functions/google-gemini/utils/client.ts
