+++ examples/google-gemini/src/functions/google-gemini/utils/index.ts
